+++ src/components/CardMenu/cardMenu.jsx
@@ -1,16 +1,18 @@
-import { Link } from "react-router";
 import styles from "./cardMenu.module.css";
+import { TabLink } from "../TabLink/tabLink";
+import { useSelector } from "react-redux";
+import { selectDishById } from "../../redux/entities/dishes/slice";
 
 export const CardMenu = ({ menu }) => {
   return (
     <div className={styles.container}>
       <div>
-        {menu.map(({ id, name }) => {
+        {menu.map((id) => {
+          const { name } = useSelector((state) => selectDishById(state, id));
+
           return (
             <div key={id} className={styles.item}>
-              <Link className={styles.text} to={`/dish/${id}`}>
-                {name}
-              </Link>
+              <TabLink to={`/dish/${id}`} text={name} />
             </div>
           );
         })}
