+++ src/components/CardMenu/cardMenu.module.css
@@ -15,14 +15,3 @@
   margin-bottom: 5px;
   padding-left: 10px;
 }
-
-.text {
-  color: inherit;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.text:hover {
-  color: #007bff;
-  text-decoration: none;
-}
