--- src/components/Greeting/greeting.module.css
@@ -1,38 +1,0 @@
-.outerContainer {
-  justify-content: center;
-  align-items: center;
-  display: flex;
-}
-
-.innerContainer {
-  text-align: center;
-  justify-content: center;
-  align-items: center;
-  padding: 20px;
-  border: 1px solid #e5e5e5;
-  border-radius: 10px;
-  background-color: white;
-  width: 600px;
-}
-
-.default {
-  margin-left: 10px;
-  margin-right: 10px;
-  display: inline-block;
-  padding: 10px 15px;
-  text-decoration: none;
-  border-radius: 5px;
-  border: none;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.lightTheme {
-  color: black;
-  background-color: #d3d3d3;
-}
-
-.darkTheme {
-  color: white;
-  background-color: #3c3c3c;
-}
