+++ src/components/Layout/layout.jsx
@@ -1,6 +1,6 @@
 import { LoginForm } from "../LoginForm/loginForm";
 import { ProgressBar } from "../ProgressBar/progressBar";
-import { ThemeConstextSwitcher } from "../ThemeContextSwitcher/themeContextSwitcher";
+import { ThemeContextSwitcher } from "../ThemeContextSwitcher/themeContextSwitcher";
 import { Cart } from "../Cart/cart";
 import styles from "./layout.module.css";
 import { Outlet } from "react-router";
@@ -13,7 +13,7 @@ export const Layout = () => {
         HEADER
         <div>
           <LoginForm />
-          <ThemeConstextSwitcher />
+          <ThemeContextSwitcher />
         </div>
       </div>
       <section>
