+++ src/components/RestaurantTab/restaurantTab.jsx
@@ -1,10 +1,8 @@
-import styles from "./restaurantTab.module.css";
 import { useSelector } from "react-redux";
 import { selectRestaurantById } from "../../redux/entities/restaurants/slice";
-import { NavLink } from "react-router";
 import { use } from "react";
-import classNames from "classnames";
 import { ThemeContext } from "../ThemeContext";
+import { TabNavLink } from "../TabNavLink/tabNavLink";
 
 export const RestaurantTab = ({ id }) => {
   const restaurant = useSelector((state) => selectRestaurantById(state, id));
@@ -16,18 +14,5 @@ export const RestaurantTab = ({ id }) => {
 
   const { name } = restaurant;
 
-  return (
-    <NavLink
-      className={({ isActive }) =>
-        classNames(styles.default, {
-          [styles.lightTheme]: theme === "light",
-          [styles.darkTheme]: theme === "dark",
-          [styles.isActive]: isActive,
-        })
-      }
-      to={id}
-    >
-      {name}
-    </NavLink>
-  );
+  return <TabNavLink to={id} text={name} />;
 };
