+++ src/components/TabLink/tabLink.module.css
@@ -4,26 +4,23 @@
   display: inline-block;
   padding: 10px 15px;
   text-decoration: none;
-  border-radius: 5px;
   border: none;
   cursor: pointer;
   transition: background-color 0.3s;
+  color: inherit;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.default:hover {
+  color: #007bff;
+  text-decoration: none;
 }
 
 .lightTheme {
   color: black;
-  background-color: #d3d3d3;
 }
 
 .darkTheme {
-  color: white;
-  background-color: #3c3c3c;
-}
-
-.isActive.lightTheme {
-  background-color: #f3f3f3;
-}
-
-.isActive.darkTheme {
-  background-color: #aeaeae;
+  color: #919191;
 }
