--- src/pages/homePage.jsx
@@ -1,5 +1,0 @@
-import { Greeting } from "../components/Greeting/greeting";
-
-export const HomePage = () => {
-  return <Greeting />;
-};
