+++ src/pages/homePage/homePage.module.css
@@ -1,0 +1,16 @@
+.outerContainer {
+  justify-content: center;
+  align-items: center;
+  display: flex;
+}
+
+.innerContainer {
+  text-align: center;
+  justify-content: center;
+  align-items: center;
+  padding: 20px;
+  border: 1px solid #e5e5e5;
+  border-radius: 10px;
+  background-color: white;
+  width: 600px;
+}
